requirements.txt
google-cloud-bigquery>=3.0.0
pandas>=2.0.0
pyarrow>=10.0.0
pysolr>=3.9.0
praw>=7.7.0
apache-beam[gcp]
google-cloud-bigquery-storage
torch
transformers
scikit-learn
numpy
pyyaml
streamlit>=1.28.0
sentence-transformers
onnxruntime
onnx
flask>=2.3.0
